Move some generates at the correct place
